Updated the KB sample readme file.
