--- Sample-WPF.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25420.1
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sample-WPF", "Sample-WPF\Sample-WPF.csproj", "{36F26733-E90C-436A-9C63-F416942CCCC2}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-		Release-Xml|Any CPU = Release-Xml|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{36F26733-E90C-436A-9C63-F416942CCCC2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{36F26733-E90C-436A-9C63-F416942CCCC2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{36F26733-E90C-436A-9C63-F416942CCCC2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{36F26733-E90C-436A-9C63-F416942CCCC2}.Release|Any CPU.Build.0 = Release|Any CPU
-		{36F26733-E90C-436A-9C63-F416942CCCC2}.Release-Xml|Any CPU.ActiveCfg = Release|Any CPU
-		{36F26733-E90C-436A-9C63-F416942CCCC2}.Release-Xml|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
