--- Sample-WPF/MainWindow.xaml.cs
@@ -1,32 +1,0 @@
-﻿using Syncfusion.UI.Xaml.Charts;
-using System;
-using System.Collections.Generic;
-using System.Collections.ObjectModel;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
-
-namespace Sample_WPF
-{
-    /// <summary>
-    /// Interaction logic for MainWindow.xaml
-    /// </summary>
-    public partial class MainWindow : UserControl
-    {      
-        public MainWindow()
-        {
-            InitializeComponent();                     
-        }
-      
-    }   
-
-}
