--- Sample-WPF/Model.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Sample_WPF
-{
-    public class Sales
-    {
-        public string Product { get; set; }
-
-        public double Revenue { get; set; }
-    }
-}
