--- Sample-WPF/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Syncfusion.Licensing" version="17.4.0.50" targetFramework="net46" />
-  <package id="Syncfusion.SfChart.WPF" version="17.4.0.50" targetFramework="net46" />
-</packages>
