+++ Sample_WPF/App.xaml
