+++ Sample_WPF/MainWindow.xaml
@@ -1,12 +1,12 @@
-﻿<UserControl x:Class="Sample_WPF.MainWindow"
+﻿<Window x:Class="Sample_WPF.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:Sample_WPF"
         xmlns:Syncfusion="clr-namespace:Syncfusion.UI.Xaml.Charts;assembly=Syncfusion.SfChart.WPF"
+        xmlns:local="clr-namespace:Sample_WPF"
         mc:Ignorable="d"
-        Height="600" Width="600">
+        Title="MainWindow" Height="450" Width="800">
 
     <Syncfusion:SfChart3D x:Name="Chart"  Header="3D Bar Chart ">
         <Syncfusion:SfChart3D.DataContext>
@@ -26,5 +26,6 @@
             </Syncfusion:BarSeries3D.AdornmentsInfo>
         </Syncfusion:BarSeries3D>
     </Syncfusion:SfChart3D >
-    
-</UserControl>
+
+
+</Window>
