+++ Sample_WPF/MainWindow.xaml.cs
@@ -1,0 +1,15 @@
+﻿using System.Windows;
+
+namespace Sample_WPF
+{
+    /// <summary>
+    /// Interaction logic for MainWindow.xaml
+    /// </summary>
+    public partial class MainWindow : Window
+    {
+        public MainWindow()
+        {
+            InitializeComponent();
+        }
+    }
+}
