+++ Sample_WPF/Model.cs
@@ -1,0 +1,9 @@
+﻿namespace Sample_WPF
+{
+    public class Sales
+    {
+        public string Product { get; set; }
+
+        public double Revenue { get; set; }
+    }
+}
