+++ Sample_WPF/Sample_WPF.sln
@@ -1,0 +1,22 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.12.35527.113 d17.12
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sample_WPF", "Sample_WPF.csproj", "{BCBC4B96-1651-4F18-A0DC-E23D0E9F23AB}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{BCBC4B96-1651-4F18-A0DC-E23D0E9F23AB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BCBC4B96-1651-4F18-A0DC-E23D0E9F23AB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BCBC4B96-1651-4F18-A0DC-E23D0E9F23AB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BCBC4B96-1651-4F18-A0DC-E23D0E9F23AB}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
